+++ HISTORY.rst
@@ -10,6 +10,8 @@ Changes
 * ModelicaEnv:
     - Introduced action clipping
     - model_params: None values are not passed to the OpenModelica env to allow initialization
+    - Introduced abort reward in env if episode is terminated
+
 
 Add
 ^^^