+++ HISTORY.rst
@@ -10,6 +10,7 @@ Changes
 * ModelicaEnv:
     - Introduced action clipping
     - model_params: None values are not passed to the OpenModelica env to allow initialization
+    - model_params: negative time values are introduced for initialization (fix)
     - Introduced abort reward in env if episode is terminated
     - Introduced obs_output to define a subset of history given as observation to the agent
 
@@ -18,6 +19,7 @@ Add
 ^^^
 * Random Process wrapper
 * ObsTempl test
+* reset test for initialized env